Completed features of TaskQueue, overhauled Task
TaskQueue:
TaskQueue is now feature-complete and awaits further testing.

Documentation and a specialization for the result type void is still missing.

All the accumulation functions were moved into their own namespace.

Task:
The entire state system was removed (it was causing problems in the task queue and was not used by any other parts of the thread manager).

Added moveResult() which returns a r-value reference to the result of the task.