Finalized the TaskQueue, improved other components of the thread system
TaskQueue: A fix for synchronization issues with getResult() was implemented.

TaskQueueAccumulators: NOU_FUNC was added to all of them.

Task.hpp: makeMemFuncTask()'s definition was removed (the declaration was already removed before).

ConditionVariable: NOU_CLASS has been added.

Threads.hpp: The file was added. It provides all of the components of the thread system in a single header (for easier usage by a user).