Fixed errors and improved compatibility with Clang
Fixed an error and a bunch of warnings (Some warnings are still missing
and an issue on GitHub was created [Issue #17]).

Also improved compiler detection when Clang is used.

Verfied that NOU_COMPILER, NOU_OS, NOU_OS_LIBRARY and
NOU_COMPILER_VERSION work on Linux (Ubuntu) when GCC or Clang is used.