Implemented a lot of fixes that became aparent with GCC
HashMap: remove() now does not remove a local variable reference
anymore. Also, there were previously some methods left that took
parameter by value instead of reference. This is now fixed.

StringView: Removed constexpr from almost all methods (only a few static
ones are left). This is because it turned out that it is not possible
to make StringView constexpr with the way that it is currently writte
(it has a member that is a pointer to a pointer, which is generally not
possible). For some reason this was accepted by MSVC but the errors
became aparent when compiling with GCC. Also, the unit tests have been
changed to not use constexpr StringViews anymore.

Vector: "data() const" now returns "const T*" instead of "T* const &"
the return type of the non-const variant was not changed.

mem_mngt/Utils: alignedAlloc() now works on POSIX systems. Previously,
it would almost always return a nullptr and fail to allocate a proper
chunk of memory.

UnitTests: See StringView.