Update CMakeLists.txt, fixes for Linux
The CMakeLists.txt now set the RPATH variable and link against the CMake
pacakge "Threads"(required for pthreads). This change has yet to be
tested on Windows.

For some reason, the Unittests were spilt up into to directories
("UnitTests/" and "Unittests/") which did not become apparent on
Windows but caused errors on Linux. This is fixed now.