Added C++14 support, improved convention conformance.
On Linux and under the GCC, the project can now be compiled with the
C++14 standard. Other platforms and compilers (Clang on Linux) still
require testing. The meta-function InvokeResult needed to be rewritten
for C++14 and it is not verified yet whether the C++14 version behaves
exactly like the C++17 version.

In the case that the project is compiled using C++14, the meta-functions
IsInvocable and IsInvocable_r are now disabled and a system was put in
place to check for their existance (certain macros will be defined in
the case that they do exist).

Also there are now macros to check for the current C++ standard (these
are nicer wrappers for __cplusplus and the old values of that macro)

To improve convention conformance, NOU::NOU_CORE::remove_reference and
NOU::NOU_CORE::remove_reference_t have been renamed to
NOU::NOU_CORE::RemoveReference and NOU::NOU_CORE::RemoveReference_t
respectively.