Fixed compile errors, enhanced compatiblilty with Clang
Fixed an error in the GPA where some methods had a default value for the
first parameter but not for the second.

Clang should now be properly identified by the library, and the CMake
commands that would disable certain defines for Clang (those defines
that were supposed to deactivate defines like __GNUC__) were removed.

The CMakeLists.txt now has an option to enable C++14 partial compatiblilty
and the README.md was updated to informat about that mode.