Fully implemented C++14 compability mode
The C++14 compability mode is now implemented and works for both Clang
and GCC on Linux.

The original system to identify the C++ standard has been removed and
replaced with a "switch" (macro that is either defined or not) to check
whether the library needs to be compiled in compability mode or not. The
original system did not work with MSVC.

All static constexpr members are now also defined out-of-class instead
of only being declared within the class.